Añadir modelos necesarios para tickets

+++ src/Models/Factura/Factura.ts
@@ -30,7 +30,7 @@ Factura.init({
         defaultValue: false
     },
     total: {
-      type: Sequelize.FLOAT(10,2),
+      type: Sequelize.FLOAT(12,2),
       allowNull: false
     }
 }, {
